Modifica dei permessi relativi ai posti e alla modifica del profilo.
Aggiunta dei metodi getAll users, validatorUser, update user

+++ SocialNetworkBE/src/main/java/com/SCAI/socialBan/service/PostServiceImplementation.java
@@ -14,10 +14,13 @@ import com.SCAI.socialBan.model.User;
 import com.SCAI.socialBan.repository.PostRepository;
 
 @Service
-public class PostServiceImplementation implements PostService{
+public class PostServiceImplementation implements PostService {
 
 	@Autowired
 	private PostRepository postRepository;
+
+	@Autowired
+	private UserService userService;
 
 	@Override
 	public Post createPost(Post post, User user) {
@@ -29,8 +32,7 @@ public class PostServiceImplementation implements PostService{
 		createdPost.setUser(user);
 		createdPost.setCreatedAt(LocalDateTime.now());
 		createdPost.setComments(new ArrayList<>());
-		
-		
+
 		return postRepository.save(createdPost);
 	}
 
@@ -38,27 +40,46 @@ public class PostServiceImplementation implements PostService{
 	public Post findPostById(Long id) throws Exception {
 		Optional<Post> opt = postRepository.findById(id);
 
-		if(opt.isPresent()) 
-		{
+		if (opt.isPresent()) {
 			return opt.get();
 		}
 
-		throw new Exception("post not found with id "+ id);
+		throw new Exception("post not found with id " + id);
+
 	}
 
 	@Override
-	public void deletePost(Long id) throws Exception {
-		//Il metodo di deleteById(id)  se lo trova lo elimina altrimenti ignora, perciÃƒÂ² non dovrebbe servire il find.
-		//findPostById(id);
-		
+	public void deletePost(Long id, String jwt) throws Exception {
+		// Il metodo di deleteById(id) se lo trova lo elimina altrimenti ignora,
+		// perciÃƒÂ² non dovrebbe servire il find.
+		// findPostById(id);
+		Post post = findPostById(id);
+		User postOwner = userService.findUserById(post.getUser().getId());
+		if(!userService.verificationUser(postOwner,jwt)){
+			throw new Exception("Utente non autorizzato all'eliminazione ");
+		}
+
 		postRepository.deleteById(id);
 	}
 
 	@Override
-	public Post updatePost(Post post, Long id) throws Exception {
+	public Post updatePost(Post post, Long id, String jwt) throws Exception {
 		
 		//old post to update
 		Post oldPost = findPostById(id);
+		User userPostOwner = userService.findUserById(oldPost.getUser().getId());
+
+		if(!userService.verificationUser(userPostOwner,jwt)){
+			throw new Exception("Utente non autorizzato alla modifica ");
+					// + "USER PROPRIETARIO     "
+					// + userPostOwner.getFullName() + "    |   "
+		 			// + userPostOwner.getId() + "    |    "
+		 			// + userPostOwner.getEmail()
+					// + "USER MODIFICANTE      "
+					// + userPostEditing.getFullName() + "    |   "
+					// + userPostEditing.getId() + "    |    "
+		 			// + userPostEditing.getEmail());
+		}
 		
 		//updating all post attributes
 		if(post.getTitle() != null) {
@@ -84,12 +105,12 @@ public class PostServiceImplementation implements PostService{
 	@Override
 	public Post likePost(Long postId, User user) throws Exception {
 		Post post = findPostById(postId);
-		
-		//se la lista di likes contiene l'id dello user, lo rimuove, altrimenti lo aggiunge
-		if(post.getLikes().contains(user.getId())){
+
+		// se la lista di likes contiene l'id dello user, lo rimuove, altrimenti lo
+		// aggiunge
+		if (post.getLikes().contains(user.getId())) {
 			post.getLikes().remove(user.getId());
-		}
-		else {
+		} else {
 			post.getLikes().add(user.getId());
 		}
 		return postRepository.save(post);
@@ -98,17 +119,17 @@ public class PostServiceImplementation implements PostService{
 	@Override
 	public Post commentPost(Long postId, User user, String content) throws Exception {
 		Post post = findPostById(postId);
-		
-		if(post == null) {
+
+		if (post == null) {
 			throw new Exception("Post Id not valid, post not found " + postId);
 		}
-		
+
 		Comment comment = new Comment();
 		comment.setPostId(postId);
 		comment.setUser(user);
 		comment.setText(content);
 		comment.setCreatedAt(LocalDateTime.now());
-	
+
 		post.getComments().add(comment);
 
 		return postRepository.save(post);
